Split view and metric view create tests into dedicated files
Move the metric-view cases out of test_table.py into test_metric_view.py
and add test_view.py so both view and metric view cover optional columns.
